actor& bulletのゲッターを intからfloatに

+++ Cont/Shootingood/Actor.cpp
@@ -1,20 +1,20 @@
 #include "Actor.h"
 
 // 現在のx座標を返す
-int Actor::Get_X() {
+float Actor::Get_X() {
 	return vector3->x;
 }
 // x座標をセットする
-void Actor::Set_X(int set_x) {
+void Actor::Set_X(float set_x) {
 	vector3->x = set_x;
 }
 
 // 現在のy座標を返す
-int Actor::Get_Y() {
+float Actor::Get_Y() {
 	return vector3->y;
 }
 // y座標を設定する
-void Actor::Set_Y(int set_y) {
+void Actor::Set_Y(float set_y) {
 	vector3->y = set_y;
 }
 // 幅を返す
